throw error async to trigger uncaughtException, update tests
