+++ lib/ContentTypeParser.js
@@ -51,7 +51,7 @@ ContentTypeParser.prototype.run = function (contentType, handler, request, reply
   var result = this.getHandler(contentType)(request.req, done)
   if (result && typeof result.then === 'function') {
     result.then(body => done(null, body)).catch(err => {
-      handlePromiseError(err, reply)
+      if (handlePromiseError(err, reply)) return
       done(err, null)
     })
   }
