+++ lib/handleRequest.js
@@ -119,7 +119,7 @@ function preHandlerCallback (err, reply) {
         reply.send(payload)
       }
     }).catch((err) => {
-      handlePromiseError(err, reply)
+      if (handlePromiseError(err, reply)) return
       reply._isError = true
       reply.send(err)
     })
