+++ lib/reply.js
@@ -207,7 +207,7 @@ function handleError (reply, error, cb) {
     if (result && typeof result.then === 'function') {
       result.then(payload => reply.send(payload))
             .catch(err => {
-              handlePromiseError(err, reply)
+              if (handlePromiseError(err, reply)) return
               reply.send(err)
             })
     }
