+++ test/custom-parser-system-errors.test.js
@@ -39,12 +39,12 @@ fastify.listen(0, err => {
 
   // Tap patched this event and we have no chance to listen on it.
   // Since any test runs in a seperate process we don't have to restore it
-  process.removeAllListeners('unhandledRejection')
+  process.removeAllListeners('uncaughtException')
 
   test('Should exit due to EvalError', t => {
     t.plan(2)
 
-    process.once('unhandledRejection', (err) => {
+    process.once('uncaughtException', (err) => {
       req.abort()
       t.type(err, systemErrors.EvalError)
     })
@@ -64,7 +64,7 @@ fastify.listen(0, err => {
   test('Should exit due to RangeError', t => {
     t.plan(2)
 
-    process.once('unhandledRejection', (err) => {
+    process.once('uncaughtException', (err) => {
       req.abort()
       t.type(err, systemErrors.RangeError)
     })
@@ -84,7 +84,7 @@ fastify.listen(0, err => {
   test('Should exit due to ReferenceError', t => {
     t.plan(2)
 
-    process.once('unhandledRejection', (err) => {
+    process.once('uncaughtException', (err) => {
       req.abort()
       t.type(err, systemErrors.ReferenceError)
     })
@@ -104,7 +104,7 @@ fastify.listen(0, err => {
   test('Should exit due to SyntaxError', t => {
     t.plan(2)
 
-    process.once('unhandledRejection', (err) => {
+    process.once('uncaughtException', (err) => {
       req.abort()
       t.type(err, systemErrors.SyntaxError)
     })
@@ -124,7 +124,7 @@ fastify.listen(0, err => {
   test('Should exit due to TypeError', t => {
     t.plan(2)
 
-    process.once('unhandledRejection', (err) => {
+    process.once('uncaughtException', (err) => {
       req.abort()
       t.type(err, systemErrors.TypeError)
     })
@@ -144,7 +144,7 @@ fastify.listen(0, err => {
   test('Should exit due to AssertionError', t => {
     t.plan(2)
 
-    process.once('unhandledRejection', (err) => {
+    process.once('uncaughtException', (err) => {
       req.abort()
       t.type(err, systemErrors.AssertionError)
     })
